moved usercontroller and user route to patient controller and patient route
